Add files via upload
The primary goal of an e-commerce site is to sell goods online. This project deals with developing an e-commerce website for Online Product Sale. It provides the user with a catalog of different product available for purchase in the store. In order to facilitate online purchase a shopping cart is provided to the user.